+++ lib/src/main/kotlin/nl/mirrajabi/humanize/duration/languages/LanguageDictionary.kt
@@ -1,5 +1,5 @@
 package nl.mirrajabi.humanize.duration.languages
 
 interface LanguageDictionary {
-    fun get(key: String, count: Float): String
+    fun get(key: String, count: Long): String
 }
